created a dynamic page routing added dashboard to layout.tsx adjusted width of dashboard to fit
